# services/gateway-api/openapi.yaml
openapi: "3.1.0"
info:
  title: Opencomplai Gateway API
  version: 0.9.0
  description: |
    REST API gateway for the Opencomplai compliance platform.
    All routes proxy to the appropriate Python service.
    This spec is the source of truth for the API contract.

    Manifests can name several frameworks (compliance_targets). Risk
    classification, dossier generation and the risk engine's public
    /v1/checker/* routes (called by the docs widget, not proxied here) are
    EU AI Act only.

    Authentication: every /v1 route requires credentials, either an
    x-api-key header (API-key mode) or an Authorization bearer JWT (OIDC
    mode). A deployment uses one of the two. Only GET /health is open. The
    gateway refuses to start unconfigured unless authentication is explicitly
    disabled for local development (OPENCOMPLAI_AUTH_DISABLED=1).
servers:
  - url: http://localhost:8080
    description: Local Docker Compose deployment
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /v1/sync/metadata:
    post:
      operationId: syncMetadata
      summary: Sync allowlisted metadata through egress proxy (idempotent)
      description: >-
        The egress proxy checks the body against its field allowlist and
        refuses the whole request with 403 EGRESS_BLOCKED if any other key is
        present. When no dashboard is configured (PRO_DASHBOARD_URL unset)
        nothing is sent anywhere: the proxy only confirms that the payload is
        allowed. When one is configured, the proxy forwards the payload to it
        and returns the dashboard's own response and status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncMetadataRequest"
      responses:
        "200":
          description: >-
            Payload validated. The body shown is the proxy's own confirmation,
            returned when no dashboard is configured; with a dashboard
            configured the body is the dashboard's response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncMetadataResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: >-
            A field outside the allowlist, or a dashboard URL that is not an
            allowed destination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EgressBlockedError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: >-
        API-key mode, for self-hosted single-operator deployments: the value
        of OPENCOMPLAI_API_KEY. A shared key authenticates the deployment as a
        whole, as the tenant named by OPENCOMPLAI_API_KEY_TENANT_ID (default
        oss-default). Ignored when OIDC_JWKS_URI is set. See
        docs/src/deployment/authentication.md.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        OIDC mode, for multi-user deployments (OIDC_JWKS_URI set): an RS256
        JWT verified against the identity provider's JWKS, with iss, aud, exp
        and nbf enforced. The token must also carry the tenant claim named by
        OIDC_TENANT_CLAIM. See docs/src/deployment/authentication.md.
  headers:
...
    RetryAfter:
      description: Seconds to wait before retrying; sent only with the rate limiter's own 429
      schema:
        type: integer
...
  responses:
    Unauthorized:
      description: >-
        Credentials missing or invalid (POLICY_DENIED). The message says
        which: "Invalid or missing API key." in API-key mode; "Missing Bearer
        token.", "Invalid or expired JWT." or "Token missing tenant claim." in
        OIDC mode.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: >-
        Too many requests. After 10 failed authentication attempts from one
        client within 60 seconds (the defaults) further failures return 429
        RATE_LIMITED in the gateway envelope instead of 401. Separately, more
        than 300 requests of any kind from one client within 60 seconds (the
        default; the Docker Compose stack raises it to 1000) return the rate
        limiter's own body, {statusCode, error, message}, with a Retry-After
        header. Both limits are configurable.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ErrorEnvelope"
              - $ref: "#/components/schemas/RequestRateLimitError"
...
    DependencyUnavailable:
      description: An upstream service could not be reached
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DependencyUnavailableError"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error_code, message, category, retryable, correlation_id]
      properties:
        error_code:
          type: string
        message:
          type: string
        category:
          type: string
          enum: [client, policy, system, dependency]
        retryable:
          type: boolean
        correlation_id:
          type: string
        remediation_hint:
          type: string
    RequestRateLimitError:
      type: object
      required: [statusCode, error, message]
      description: >-
        Body the gateway's request rate limiter sends when a client exceeds
        the global limit. It does not use the ErrorEnvelope shape.
      properties:
        statusCode:
          type: integer
          enum: [429]
        error:
          type: string
          enum: [Too Many Requests]
        message:
          type: string
          description: Rate limit exceeded, retry in <duration>
...
    DependencyUnavailableError:
      type: object
      required: [error_code, message, retryable]
      description: >-
        Returned with 503 when an upstream service cannot be reached. The
        gateway adds category and correlation_id; the egress proxy's variant
        carries only error_code, message and retryable.
      properties:
        error_code:
          type: string
          enum: [DEPENDENCY_UNAVAILABLE]
        message:
          type: string
        category:
          type: string
          enum: [dependency]
        retryable:
          type: boolean
        correlation_id:
          type: string
...
    SyncMetadataRequest:
      type: object
      description: >-
        Only the fields listed here may be sent; any other key is refused with
        403 EGRESS_BLOCKED. The egress proxy checks field names only, not
        their types, and requires none of them.
      additionalProperties: false
      properties:
        system_id:
          type: string
        commit_ref:
          type: string
        policy_bundle_version:
          type: string
        risk_class:
          type: string
        control_pass_rate:
          type: number
        control_fail_rate:
          type: number
        pending_verifications_count:
          type: integer
        bundle_checksum:
          type: string
        size_bytes:
          type: integer
        signed_by:
          type: string
        timestamp:
          type: string
        pass_count:
          type: integer
        fail_count:
          type: integer
        trap_frequency:
          type: number
        override_rate:
          type: number
        eval_set_id: {}
        eval_overall_outcome: {}
        eval_failed_evaluator_ids: {}
    SyncMetadataResponse:
      type: object
      required: [status, destination, fields_validated, field_count]
      properties:
        status:
          type: string
          enum: [synced]
        destination:
          type: "null"
          description: Always null; no dashboard was contacted
        fields_validated:
          type: boolean
          enum: [true]
        field_count:
          type: integer
          description: Number of fields in the request body
    EgressBlockedError:
      type: object
      required: [error_code, message, category, retryable, policy_hash]
      description: >-
        Returned by the egress proxy with 403. It is not wrapped in detail and
        carries no correlation_id.
      properties:
        error_code:
          type: string
          enum: [EGRESS_BLOCKED]
        message:
          type: string
        category:
          type: string
          enum: [policy]
        retryable:
          type: boolean
        policy_hash:
          type: string
          description: sha256:<hex> of the allowlist configuration that was in force
...